requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
multidict==6.0.2
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
secedgar==0.4.0a3
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
